services/shared/pyproject.toml
[tool.poetry]
name = "carpeta-ciudadana-shared"
version = "0.1.0"
description = "Shared utilities for Carpeta Ciudadana services"
authors = ["Carpeta Ciudadana Team"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# AWS SDK (legacy)
boto3 = "^1.34.0"

# Azure SDK (nuevo)
azure-identity = "^1.15.0"
azure-storage-blob = "^12.19.0"
azure-servicebus = "^7.11.4"
azure-keyvault-certificates = "^4.7.0"
azure-keyvault-secrets = "^4.7.0"

# Database
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"

# Search (compatible con ambos)
opensearch-py = "^2.4.0"

# Other
redis = "^5.0.1"
httpx = "^0.26.0"
cryptography = "^41.0.7"
python-jose = "^3.3.0"
jwcrypto = "^1.5.0"

# OpenTelemetry
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"
opentelemetry-instrumentation-httpx = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
ruff = "^0.1.9"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
